rename tests and move functions
