+++ hhnk_research_tools/__init__.py
@@ -1,4 +1,11 @@
-from hhnk_research_tools.gis.raster import Raster, RasterMetadata
+# hhnk_research_tools/__init__.py
+from typing import TYPE_CHECKING
+
+from hhnk_research_tools.gis.raster import Raster, RasterMetadata  # noqa: F401
+
+if TYPE_CHECKING:
+    import hhnk_research_tools as hrt
+
 import hhnk_research_tools.threedi as threedi
 import hhnk_research_tools.variables as variables
 import hhnk_research_tools.waterschadeschatter.resources
@@ -36,11 +43,9 @@ from hhnk_research_tools.general_functions import (
     load_source,
     time_delta,
 )
-
-from hhnk_research_tools.gis.raster_blocks import RasterBlocks
+from hhnk_research_tools.gis.raster_calculator import RasterBlocks, RasterCalculatorV2
 from hhnk_research_tools.raster_functions import (
     RasterCalculator,
-    RasterCalculatorV2,
     build_vrt,
     create_meta_from_gdf,
     create_new_raster_file,
