+++ hhnk_research_tools/raster_functions.py
@@ -7,7 +7,6 @@ import numpy as np
 from IPython.display import display
 from osgeo import gdal, ogr
 
-import hhnk_research_tools as hrt
 from hhnk_research_tools.folder_file_classes.folder_file_classes import Folder
 from hhnk_research_tools.general_functions import (
     check_create_new_file,
@@ -529,147 +528,3 @@ def hist_stats(histogram: dict, stat_type: str, ignore_keys=[0]):
             total += histogram[value]
             if total >= median_index:
                 return value
-
-
-class RasterCalculatorV2:
-    def __init__(
-        self,
-        raster_out: Raster,
-        raster_paths_dict: dict,
-        nodata_keys: list,
-        mask_keys: list,
-        metadata_key: str,
-        custom_run_window_function: types.MethodType,
-        output_nodata: int = -9999,
-        min_block_size: int = 4096,
-        verbose: bool = False,
-    ):
-        """
-        Base setup for raster calculations. The input rasters defined in raster_paths_dict
-        are looped over per block. Note that all input rasters should have the same extent.
-        This can be achieved with .vrt if the original rasters do not have the same extent.
-        For each block the custom_run_window_function will be run. This always takes a
-        block as input and also returns the block. For example:
-
-        def run_dem_window(block):
-            block_out = block.blocks['dem']
-
-            #Watervlakken ophogen naar +10mNAP
-            block_out[block.blocks['watervlakken']==1] = 10
-
-            block_out[block.masks_all] = nodata
-            return block_out
-
-
-        raster_out (hrt.Raster): output raster location
-        raster_paths_dict (dict): {key:hrt.Raster} these rasters will have blocks loaded.
-        nodata_keys (list): keys to check if all values are nodata, if yes then skip
-        mask_keys (list): keys to add to nodatamask
-        metadata_key (str): key in raster_paths_dict that will be used to
-            create blocks and metadata
-        custom_run_window_function: function that does calculation with blocks.
-            function takes block (hrt.RasterBlocks) and kwargs as input and must return block
-        output_nodata (int): nodata
-        min_block_size (int): min block size for generator blocks_df
-        verbose (bool): print progress
-        """
-        self.raster_out = raster_out
-        self.raster_paths_dict = raster_paths_dict
-        self.nodata_keys = nodata_keys
-        self.mask_keys = mask_keys
-        self.metadata_key = metadata_key
-        self.custom_run_window_function = custom_run_window_function
-        self.output_nodata = output_nodata
-        self.min_block_size = min_block_size
-        self.verbose = verbose
-
-    @property
-    def metadata_raster(self):
-        return self.raster_paths_dict[self.metadata_key]
-
-    def verify(self, overwrite) -> bool:
-        cont = True
-
-        # Check if all input rasters have the same bounds
-        bounds = {}
-        for key, r in self.raster_paths_dict.items():
-            if cont:
-                if not r.exists():
-                    print(f"Missing input raster key: {key} @ {r}")
-                    cont = False
-                    continue
-                bounds[key] = r.metadata.bounds
-
-        first_val = list(bounds.values())[0]
-        for val in bounds.values():
-            if val != first_val:
-                cont = False
-                print(f"input rasters dont have the same bounds:")
-                display(bounds)
-                break
-
-        # Check if we should create new file
-        if cont:
-            cont = hrt.check_create_new_file(output_file=self.raster_out, overwrite=overwrite)
-            if cont == False:
-                if self.verbose:
-                    print(f"output raster already exists: {self.raster_out.name} @ {self.raster_out.path}")
-
-        return cont
-
-    def create(self) -> bool:
-        """Create empty output raster
-        returns bool wether the rest of the function should continue"""
-
-        # Check if function should continue.
-
-        if self.verbose:
-            print(f"creating output raster: {self.raster_out.name} @ {self.raster_out.path}")
-        self.raster_out.create(metadata=self.metadata_raster.metadata, nodata=self.output_nodata)
-
-    def run(self, overwrite=False, **kwargs):
-        try:
-            cont = self.verify(overwrite=overwrite)
-            if cont:
-                self.create()
-
-            if cont:
-                if self.verbose:
-                    time_start = datetime.datetime.now()
-
-                self.metadata_raster.min_block_size = self.min_block_size
-                self.blocks_df = self.metadata_raster.generate_blocks()
-                self.blocks_total = len(self.blocks_df)
-
-                gdal_src = self.raster_out.open_gdal_source_write()
-                band_out = gdal_src.GetRasterBand(1)
-                for idx, block_row in self.blocks_df.iterrows():
-                    window = block_row["window_readarray"]
-                    block = hrt.RasterBlocks(
-                        window=window,
-                        raster_paths_dict=self.raster_paths_dict,
-                        nodata_keys=self.nodata_keys,
-                        mask_keys=self.mask_keys,
-                    )
-
-                    if block.cont:
-                        block_out = self.custom_run_window_function(block=block, **kwargs)
-                        if self.verbose:
-                            time_duration = hrt.time_delta(time_start)
-                            print(f"{idx} / {self.blocks_total} ({time_duration}s) - {self.raster_out.name}", end="\r")
-
-                        # Wegschrijven block
-                        self.raster_out.write_array(array=block_out, window=window, band=band_out)
-
-                # band_out.FlushCache()  # close file after writing, slow, needed?
-                band_out = None
-                if self.verbose:
-                    print("\nDone")
-            else:
-                if self.verbose:
-                    print(f"{self.raster_out.name} not created, .verify was false.")
-        except Exception as e:
-            band_out.FlushCache()
-            band_out = None
-            self.raster_out.unlink()
-            raise e
